Fix bug where kernel evaluation fails if no logger is set.
Improved logging mechanism in kernel. If logging is not activated,
the kernel no more write anything to the socket.
It is also possible to set the kernel logging level on the kernel side to reduce
the amount of messages writen to the logging socket. This is fairly advance
usage. Not sure if we expose or document it.