Interuptable startup
Asynchronous sessions can be interrupted when waiting for the Kernel to respond,
which currently can take a long time when licensing issue occurs.
In the future this should not be required anymore.

Kernel pool start method returns on the first started kernel, and maintain two lists of
tasks, one with all the initialisation tasks, one with the evaluation loops. This make
termination much cleaner in case some kernels were not yet initialised for example.

Kernel pool have evaluate_all sync method that takes an iterable and run all the expr
it contains using the pool.

There is also parallel_evaluate that takes a kernel path and a list of expr, that starts
a pool, and run evaluate_all, and terminate the pool.