Add options to documentation build script
Build script in docs now supports a new option to specify a target build directory
in place of _build.
This enable the somewhat standard workflow when building documentation for
github.io, which is to have an orphan branch gh-pages with the static webpages.
Usually this branch is checkout in a separate repository. The build script must be
able to target this new repo, while not messing up git files.

-all option changed to -a or --all for consistency, and no more calls `make clean`
which is too aggressive (it removes .git).

Add basic help.