Update documentation structure
The index.html page is located at the root and other pages are in docpages directory,
except those created by Sphinx.
Improve the build script to automate the build process so that it can target a directory,
specifically the one with the repository based on gh-pages branch.

This script is performing some `rm -r` and should be used with care. Except for Wolfram
internals no one should really needs it.