Import PIL would hang, this fixes it
Issue with loading pil and calling to_wl at the same time.
This errors because it tries to import something that doesn't
exist. I tried to add support using the API() method some of the
other libraries use, but couldn't get it to work. So there is likely
a more robust way, but just having the import at the top seems to only
load when needed, and work in all the cases I tried.