Add options to cloud session to cloud evaluate string as input form.
Most of the time it's more convenient to get the evaluation result than wrapping wlexpr
around each string.
This behaviour can be disabled for cases where the source of the input expression is
unknown, in which case we may want to always evaluate the serialized expr. Even if it's
probably a mistake, since evaluating a python string is a no-op.