Major refactoring of the former normalizer family of functions.
Normalizer is renamed Encoder, similarly `normalize` becomes `encode`, to better
match other library naming.
Normalizer now only refer to the user defined function that is eventually passed to
export and that is applied before the encoding to WL / WXF starts.
The serialization no more rely on a global variable `dispatch`. This instance still exists
but the API is only exposing a annotation @wolfram_encoder(*types). This annotation
deals with the relevant object under the hood.

Greatly simply and clarify the writing of extensions to custom types.

Type checking is now more consistent, and rely on MRO to find the best matching encoding
function. It increases the number of possible combinations of types, which in practice will
remain reasonably small. The slow step is cached, its cost is paid once and for all.

This commit is the first step to enabling user defined plugins.