Fix numpy missing in PIL encoder
When the PIL encoder is loaded, and numpy is missing,
error were raised.
Make sure that Numpy is loaded only when a function is
called, not when the module is. This makes clean errors,
happening when code is evaluated. It also allows to use
PIL without numpy in some cases where it's not required.