Fix update_dispatch available module heuristic
Now check for imported and built in modules available in sys.modules first.
Then also check for distribution available on the system, using pkg_resources.
This fix bug where a module is available, but not imported.
e.g after import PIL, numpy must be available but is not yet in sys.modules.