Use EvaluationData for local session
EvaluationData is wrapped around the evaluation step in a local kernel. As a result the output is similar to the one from the Cloud.
The first difference is that the evaluation result is WXF encoded so that a user defined consumer can be applied. The second is that
the messages text are rebuilt for readability. Finally messages expressions are wrapped into MessageName.
All evaluation result classes inherit from the same base class that delegate deserialization to children classes, but exposes various fields
from the parsed result dict.
Adapt some tests.