Replace force option in encoder.
The encoder dispatch option `force` is not fine enough. We need to be able to load built-in encoders without overwriting user defined encoders that have been registered before (e.g if export has not yet been loaded).
Adding keep_existing option to control whether or not existing mapping should be kept.
Adding replace_existing option to control whether or not to update current mapping.

Built in encoders are now loaded with keep_existing set to True.