Kernel controller is more robust to terminated process
When the kernel is crashing or when it quits (e.g. due to Quit[]) the kernel controller on the python side was not checking the state.
While waiting for a response (recv), regularly check for the process state.
More robust future manipulation, checking for aborted future before setting its value.